Move DownloadableResultSettings to downloader.py to avoid circular import
