+++ src/databricks/sql/cloudfetch/download_manager.py
@@ -4,28 +4,10 @@ from concurrent.futures import ThreadPoolExecutor
 from dataclasses import dataclass
 from typing import List, Union
 
-from databricks.sql.cloudfetch.downloader import ResultSetDownloadHandler
+from databricks.sql.cloudfetch.downloader import ResultSetDownloadHandler, DownloadableResultSettings
 from databricks.sql.thrift_api.TCLIService.ttypes import TSparkArrowResultLink
 
 logger = logging.getLogger(__name__)
-
-
-@dataclass
-class DownloadableResultSettings:
-    """
-    Class for settings common to each download handler.
-
-    Attributes:
-        is_lz4_compressed (bool): Whether file is expected to be lz4 compressed.
-        link_expiry_buffer_secs (int): Time in seconds to prevent download of a link before it expires. Default 0 secs.
-        download_timeout (int): Timeout for download requests. Default 60 secs.
-        max_consecutive_file_download_retries (int): Number of consecutive download retries before shutting down.
-    """
-
-    is_lz4_compressed: bool
-    link_expiry_buffer_secs: int = 0
-    download_timeout: int = 60
-    max_consecutive_file_download_retries: int = 0
 
 
 @dataclass
